# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WEB API",
        "title": "Rent Transport",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/Admin/Rent": {
            "post": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Create new rent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "AdminNewRent",
                "responses": {}
            }
        },
        "/api/Admin/Rent/:id": {
            "put": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Update the rent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "AdminUpdateRent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "rent id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/Admin/Rent/:rentId": {
            "get": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Get rent by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "AdminGetRent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "rent id",
                        "name": "rentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Delete the rent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "AdminDeleteRent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "rent id",
                        "name": "rentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/Admin/Rent/End/:rentId": {
            "post": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "End the rent",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "AdminEndRent",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "rent id",
                        "name": "rentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/Admin/Transport": {
            "get": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Get all transports",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "GetTransports",
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Admin add the transport",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "AdminAddTransport",
                "responses": {}
            }
        },
        "/api/Admin/Transport/:id": {
            "get": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Get transport by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "AdminGetTransport",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "transport id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Admin update the transport information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "AdminUpdateTransport",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "transport id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Admin delete the transport",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "AdminDeleteTransport",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "transport id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/Rent/TransportHistory/:trnaposrtId": {
            "get": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Get history of transport",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rent"
                ],
                "summary": "TranposrtHistory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "transport id",
                        "name": "transportId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/TransportHistory/:transportId": {
            "get": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Get  transport history",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "AdminGetTransportHitory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "transport id",
                        "name": "transportId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/UserHistory/:userId": {
            "get": {
                "security": [
                    {
                        "ApKeyAuth": []
                    }
                ],
                "description": "Get user history",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "AdminGetUserHistory",
                "responses": {}
            }
        }
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}